Move outbound port exceptions into their own package
