+++ src/main/kotlin/no/nb/mlt/wls/application/synqapi/synq/SynqBatchMoveItemPayload.kt
@@ -8,7 +8,7 @@ import jakarta.validation.constraints.PositiveOrZero
 import no.nb.mlt.wls.domain.model.HostName
 import no.nb.mlt.wls.domain.ports.inbound.MoveItemPayload
 import no.nb.mlt.wls.domain.ports.inbound.UpdateItem.UpdateItemPayload
-import no.nb.mlt.wls.domain.ports.outbound.ItemMovingException
+import no.nb.mlt.wls.domain.ports.outbound.exceptions.ItemMovingException
 
 @Schema(
     description = """Payload with Product/Item movement updates from the SynQ storage system.""",
