+++ src/main/kotlin/no/nb/mlt/wls/domain/ports/outbound/exceptions/ItemMovingException.kt
@@ -1,0 +1,6 @@
+package no.nb.mlt.wls.domain.ports.outbound.exceptions
+
+class ItemMovingException(
+    message: String,
+    cause: Throwable? = null
+) : RuntimeException(message, cause)
