+++ src/main/kotlin/no/nb/mlt/wls/domain/ports/outbound/exceptions/NotSupportedException.kt
@@ -1,0 +1,5 @@
+package no.nb.mlt.wls.domain.ports.outbound.exceptions
+
+class NotSupportedException(
+    message: String
+) : RuntimeException(message)
