+++ src/main/kotlin/no/nb/mlt/wls/domain/ports/outbound/exceptions/OrderUpdateException.kt
@@ -1,0 +1,6 @@
+package no.nb.mlt.wls.domain.ports.outbound.exceptions
+
+class OrderUpdateException(
+    message: String,
+    cause: Throwable? = null
+) : RuntimeException(message, cause)
