+++ src/main/kotlin/no/nb/mlt/wls/infrastructure/StorageEventProcessorAdapter.kt
@@ -11,9 +11,9 @@ import no.nb.mlt.wls.domain.ports.outbound.EmailNotifier
 import no.nb.mlt.wls.domain.ports.outbound.EventProcessor
 import no.nb.mlt.wls.domain.ports.outbound.EventRepository
 import no.nb.mlt.wls.domain.ports.outbound.ItemRepository
-import no.nb.mlt.wls.domain.ports.outbound.NotSupportedException
 import no.nb.mlt.wls.domain.ports.outbound.StatisticsService
 import no.nb.mlt.wls.domain.ports.outbound.StorageSystemFacade
+import no.nb.mlt.wls.domain.ports.outbound.exceptions.NotSupportedException
 import org.springframework.stereotype.Service
 
 private val logger = KotlinLogging.logger {}
