+++ src/main/kotlin/no/nb/mlt/wls/infrastructure/repositories/order/MongoOrderRepositoryAdapter.kt
@@ -7,8 +7,8 @@ import no.nb.mlt.wls.domain.model.HostName
 import no.nb.mlt.wls.domain.model.Order
 import no.nb.mlt.wls.domain.ports.inbound.exceptions.OrderNotFoundException
 import no.nb.mlt.wls.domain.ports.outbound.OrderRepository
-import no.nb.mlt.wls.domain.ports.outbound.OrderUpdateException
 import no.nb.mlt.wls.domain.ports.outbound.RepositoryException
+import no.nb.mlt.wls.domain.ports.outbound.exceptions.OrderUpdateException
 import no.nb.mlt.wls.infrastructure.config.TimeoutProperties
 import org.springframework.data.mongodb.repository.Query
 import org.springframework.data.mongodb.repository.ReactiveMongoRepository
